Leave STDIN of a runbook alone by default.
Now reading from STDIN in a runbook works as expected by default.
This, however, comes with a caveat in that now the entire runbook
script source is passed as an argument to 'bash -c', which means it
shows up in process list :( Another small issue with this is that errors
from within any function will be reported as coming from the
'environment' source file by bash rather than from $0.

The env var, RB_FROM_STDIN, which, if set to non-empty, will
run execute the runbook from STDIN of bash. This is the previous
behaviour, so to read from STDIN in the runbook, one will need to
redirect from $RB_STDIN.